+++ clang/utils/TableGen/ClangBuiltinsEmitter.cpp
@@ -376,8 +376,6 @@ private:
                                .Case("uint64_t", "UWi")
                                .Case("void", "v")
                                .Case("wchar_t", "w")
-                               .Case("fexcept_t", "Tx")
-                               .Case("fenv_t", "Te")
                                .Case("...", ".")
                                .Default("error");
       if (ReturnTypeVal == "error")